rename ArrayAssertionSamples to ArrayExpectationSamples
